+++ src/routes/+layout.svelte
@@ -26,21 +26,23 @@
   <a href="/super-debug">SuperDebug</a> |
   <a href="/nested">Nested</a> |
   <a href="/multiple">Multiple</a> |
-  <a href="/multiple2">Multiple 2</a>
   <br />
+  <a href="/multiple2">Multiple 2</a>
   <a href="/snapshot">Snapshot</a> |
   <a href="/dates">Dates</a> |
   <a href="/reset">Reset</a> |
   <a href="/url">URL</a> |
   <a href="/tainted">Tainted</a> |
-  <a href="/proxies">Proxies</a> |
-  <a href="/properly-nested">Properly nested</a>
+  <a href="/proxies">Proxies</a>
   <br />
+  <a href="/properly-nested">Properly nested</a>
   <a href="/files">Files</a> |
   <a href="/spa">SPA</a> |
   <a href="/nested-validation">Client-side validation</a> |
-  <a href="/tests/rex">Side-effect validation</a> |
-  <a href="/tests/superform-c">Side-effect 2</a>
+  <a href="/tests/rex">Side-effect validation</a>
+  <br />
+  <a href="/tests/superform-c">Side-effect 2</a> |
+  <a href="/tests/multiselect">Multi-select</a>
 </nav>
 
 <slot />